% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_prams_finder.R
\name{best_parmas_finder}
\alias{best_parmas_finder}
\title{Build models to obtain the model with the best parameters by updating the parameters
from the average of the parameters of the current model and the model one step before
Repeat the above step until the mean changes below the threshold}
\usage{
best_parmas_finder(
  params,
  dat,
  SEIR = FALSE,
  max_iter = 100,
  niter = 100,
  thres = 2e-05,
  pop_size = 2000,
  suppress_img = FALSE
)
}
\arguments{
\item{params}{parameter matrix to calculate avarege}

\item{dat}{the data passed into the epimodel}

\item{SEIR}{whether or not the model is an SEIR model}

\item{max_iter}{number of trials for generating models}

\item{niter}{number of iterations for the initialization step}

\item{thres}{threshold for parameters' change}

\item{pop_size}{the size of the population}

\item{suppress_img}{whether or not to supress the graph output}
}
\value{
model with the best parameters according to the algorithm
}
\description{
Build models to obtain the model with the best parameters by updating the parameters
from the average of the parameters of the current model and the model one step before
Repeat the above step until the mean changes below the threshold
}
