% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_process.R
\name{d_meas_process}
\alias{d_meas_process}
\title{Calculate the density distribution of binomial model according to the specified variable and probability}
\usage{
d_meas_process(state, meas_vars, params, log = TRUE)
}
\arguments{
\item{state}{the name of the compartments.}

\item{meas_vars}{variable to be measured, but here we will use the observed and augmented subjects, ie. I_observed and I_augmented, here.}

\item{params}{the named list of parameters}

\item{log}{whether or not to output the logarithmic result}
}
\value{
valuse of density from the specified size and probability
}
\description{
Calculate the density distribution of binomial model according to the specified variable and probability
}
