% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg.R
\name{get_avg}
\alias{get_avg}
\title{Calculate the average for each columns}
\usage{
get_avg(param)
}
\arguments{
\item{param}{parameter matrix to calculate avarege}
}
\value{
calcualte average with 10 burn-in values
}
\description{
Calculate the average for each columns
}
