% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_utils.R
\name{gibbs_kernel_SEIR}
\alias{gibbs_kernel_SEIR}
\title{Using Gibbs sampler to update parameters from the full conditional distibutions for SEIR model}
\usage{
gibbs_kernel_SEIR(epimodel)
}
\arguments{
\item{epimodel}{the class of epidemic mpdel}
}
\value{
updated epimodel
}
\description{
Using Gibbs sampler to update parameters from the full conditional distibutions for SEIR model
}
\references{
\href{https://github.com/fintzij/BDAepimodel/blob/f73daebff0d46bbd7e68af1429f37b4665fae92b/R/gibbs_template.R}{Gibbs template}
}
