% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_process.R
\name{r_meas_process}
\alias{r_meas_process}
\title{Using binomial distribution to sample values according to specified variable and probability}
\usage{
r_meas_process(state, meas_vars, params)
}
\arguments{
\item{state}{the name of the compartments.}

\item{meas_vars}{variable to be measured}

\item{params}{the named list of parameters}
}
\value{
samples from meas_vars with probabiliy from specified param
}
\description{
Using binomial distribution to sample values according to specified variable and probability
}
