% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{seir_plot}
\alias{seir_plot}
\title{Plot SIER models on confirmed cases, death cases and recovered cases.}
\usage{
seir_plot(pred, data)
}
\arguments{
\item{pred}{prediction values of confirmed cases, death cases and recovered cases. \verb{(I, D, C)}}

\item{data}{true values of confirmed cases, death cases and recovered cases. \verb{(I, D, C)}}
}
\value{
The plots of confirmed cases, death cases and recovered cases.
}
\description{
Plot SIER models on confirmed cases, death cases and recovered cases.
}
