% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sier_predict.R
\name{sier_plot}
\alias{sier_plot}
\title{Calculate quantile regerssion loss}
\usage{
sier_plot(pred, data)
}
\arguments{
\item{a1}{value of y-X*beta}

\item{a2}{value of probability}
}
\value{
z x ( tau - 0)$ if z < 0 or $ z x (tau - 1)  otherwise
}
\description{
Calculate quantile regerssion loss
}
