% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sier_predict.R
\name{sier_predict}
\alias{sier_predict}
\title{Calculate quantile regerssion loss}
\usage{
sier_predict(a1, a2, b, s, g, E, I, D, C)
}
\arguments{
\item{a1}{value of y-X*beta}

\item{a2}{value of probability}
}
\value{
z x ( tau - 0)$ if z < 0 or $ z x (tau - 1)  otherwise
}
\description{
Calculate quantile regerssion loss
}
