% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plot.R
\name{simulation_plot}
\alias{simulation_plot}
\title{Calling simulate_epimodel to simulate infected people count in 100 time steps}
\usage{
simulation_plot(epimodel, orig_data, SEIR = FALSE)
}
\arguments{
\item{epimodel}{parameter matrix to calculate avarege}

\item{orig_data}{the original data set to compare}

\item{SEIR}{whether the model is an SEIR (TRUE) or SIR(FALSE) model. Default: FALSE}
}
\value{
none, but will print the comparison plots
}
\description{
Calling simulate_epimodel to simulate infected people count in 100 time steps
}
