% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_fit}
\alias{sir_fit}
\title{Define a helper function to enforce additional constraints on beta and gamma so that
the optimal parameters will be computed until they meet the requirements}
\usage{
sir_fit(pred_days, beta, gamma, S, I, R, N)
}
\arguments{
\item{pred_days}{number of days we want to predict for.}

\item{beta}{value of alpha1, the infectious rate by exposed people.}

\item{gamma}{value of alpha1, the infectious rate by exposed people.}

\item{S}{value of alpha1, the infectious rate by exposed people.}

\item{I}{value of alpha1, the infectious rate by exposed people.}

\item{R}{value of alpha1, the infectious rate by exposed people.}

\item{N}{value of alpha1, the infectious rate by exposed people.}
}
\value{
The prediction i.e. confirmed, death, recovered based on the params for pred_days \verb{(I, D, C)}
}
\description{
Define a helper function to enforce additional constraints on beta and gamma so that
the optimal parameters will be computed until they meet the requirements
}
