% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_mle_param}
\alias{sir_mle_param}
\title{Maximize the log-likelihood function and compute initial estimates of beta and gamma}
\usage{
sir_mle_param(init_b, init_g, lower_g, upper_g, upper_b, confirmed)
}
\arguments{
\item{init_b}{Inital guess value of beta.}

\item{init_g}{Inital guess value of gamma.}

\item{lower_g}{The lower bound of gamma when using mle to optimize gamma.}

\item{upper_g}{The upper bound of gamma when using mle to optimize gamma.}

\item{upper_b}{The upper bound of beta when using mle to optimize beta.}

\item{confirmed}{A list of number of confirmed cases of real data \verb{(day_1, ..., day_N)}.}
}
\value{
The estimated parameter beta and gamma within the limit of bound values \verb{(beta, gamma)}.
}
\description{
Maximize the log-likelihood function and compute initial estimates of beta and gamma
}
\details{
Generate Yi that follows a poission distribution with a rate of I(ti).
Then derive a log-likelihood estimator and compute the initial beta and gamma by MLE.
}
