% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_ode_fit}
\alias{sir_ode_fit}
\title{Fit the SIR model using ODE to find the prediction values for susceptible, infected and removed cases
for a period of time, given gamma, beta and initial values of populations.}
\usage{
sir_ode_fit(opt_param, times, N, init_S, init_I, init_R)
}
\arguments{
\item{opt_param}{The estimated parameter beta and gamma within the limit of bound values \verb{(beta, gamma)}.}

\item{times}{A vector from 1 to the number of day we want to fit \verb{(1, 2, ..., N)}.}

\item{N}{The value of otal population.}

\item{init_S}{The value of initial susceptible people.}

\item{init_I}{The value of initial infected people.}

\item{init_R}{The value of initial removed people.}
}
\value{
The prediction i.e. susceptible, infected, removed \verb{(S, I, C)}.
}
\description{
Fit the SIR model using ODE to find the prediction values for susceptible, infected and removed cases
for a period of time, given gamma, beta and initial values of populations.
}
\details{
The reparametrized SIR model is\preformatted{S_new = S_old - beta * S_old * I_old / N
I_new = I_old + beta * S_old * I_old / N - gamma * I_old
R_new = R_old + gamma * I_old
}
}
