% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_plot}
\alias{sir_plot}
\title{Calculate quantile regerssion loss}
\usage{
sir_plot(pred, data)
}
\arguments{
\item{pred}{contact rate}

\item{data}{removed rate}
}
\description{
Calculate quantile regerssion loss
}
