% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seird_m.R
\name{seird_m_fit}
\alias{seird_m_fit}
\title{Fit SEIR model}
\usage{
seird_m_fit(
  pred_days,
  a1,
  a2,
  b,
  s,
  g,
  E,
  I,
  D,
  C,
  a1_dec_rate = 1,
  a2_dec_rate = 1
)
}
\arguments{
\item{pred_days}{number of days we want to predict for.}

\item{a1}{value of alpha1, the infectious rate by exposed people.}

\item{a2}{value of alpha2, the infectious rate by infected people.}

\item{b}{value of beta, the rate of transition from exposed to infected.}

\item{s}{value of sigma, the rate of transition from infected to recovered.}

\item{g}{value of gmma, the rate of transition from infected to death.}

\item{E}{initial number of infected people with no symptom.}

\item{I}{initial number of confirmed infected people with symptom.}

\item{D}{initial number of death.}

\item{C}{initial number of recovered people.}

\item{a1_dec_rate}{the rate of decreasing alpha1, default is 1.}

\item{a2_dec_rate}{the rate of decreasing alpha2, default is 1.}
}
\value{
The prediction i.e. confirmed, death, recovered based on the params for pred_days \verb{(I, D, C)}.
}
\description{
Fit SEIR model
}
\details{
The reparametrized SEIRD modified model is\preformatted{E = E + alpha1 * E + alpha2 * I - beta * E
I <- I + beta * E - sigma * I - gamma * I
D <- D + gamma * I
C <- C + sigma * I
}

where \code{alpha1} and \code{alpha2} can be given a constant decaying rate by a1_dec_rate and a2_dec_rate respectively.
}
