% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_fit}
\alias{sir_fit}
\title{Fit SIR model using an iteration approach}
\usage{
sir_fit(pred_days, beta, gamma, S, I, R, N)
}
\arguments{
\item{pred_days}{Number of days we want to predict for.}

\item{beta}{Value of beta, the rate of transition from susceptible to infected.}

\item{gamma}{Value of gamma, the rate of transition from infected to removed.}

\item{S}{Initial number of susceptible people.}

\item{I}{Initial number of confirmed infected people.}

\item{R}{Initial number of removed cases.}

\item{N}{Total population.}
}
\value{
The prediction i.e. confirmed, death, recovered based on the params for pred_days \verb{(I, D, C)}
}
\description{
Fit SIR model using an iteration approach
}
\details{
The reparametrized SEIRD modified model is\preformatted{S_new = S_old - beta * S_old * I_old / N
I_new = I_old + beta * S_old * I_old / N - gamma * I_old
R_new = R_old + gamma * I_old
}
}
