% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_mle_param}
\alias{sir_mle_param}
\title{Maximize the log-likelihood function and compute initial estimates of beta and gamma}
\usage{
sir_mle_param(init_b, init_g, lower_g, upper_g, upper_b)
}
\arguments{
\item{init_b}{number of days we want to predict for.}

\item{init_g}{value of alpha1, the infectious rate by exposed people.}
}
\value{
The prediction i.e. confirmed, death, recovered based on the params for pred_days \verb{(I, D, C)}
}
\description{
Maximize the log-likelihood function and compute initial estimates of beta and gamma
}
