% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_ode_fit}
\alias{sir_ode_fit}
\title{Define a helper function to enforce additional constraints on beta and gamma so that
the optimal parameters will be computed until they meet the requirements}
\usage{
sir_ode_fit(opt_param, times, N, init_S, init_I, init_R)
}
\arguments{
\item{opt_param}{number of days we want to predict for.}

\item{times}{number of days we want to predict for.}

\item{N}{value of alpha1, the infectious rate by exposed people.}

\item{init_S}{value of alpha1, the infectious rate by exposed people.}

\item{init_I}{value of alpha1, the infectious rate by exposed people.}

\item{init_R}{value of alpha1, the infectious rate by exposed people.}
}
\value{
The prediction i.e. confirmed, death, recovered based on the params for pred_days \verb{(I, D, C)}
}
\description{
Define a helper function to enforce additional constraints on beta and gamma so that
the optimal parameters will be computed until they meet the requirements
}
