% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_plot}
\alias{sir_plot}
\title{Plot SIR models on confirmed cases and removed cases.}
\usage{
sir_plot(pred, pred_ode, data)
}
\arguments{
\item{pred}{Iteration prediction values of confirmed cases and removed cases. \verb{(I, R)}.}

\item{pred_ode}{ODE prediction values of confirmed cases and removed cases. \verb{(I, R)}.}

\item{data}{True values of confirmed cases and removed cases. \verb{(I, R)}.}
}
\value{
Two plots of infection cases and cremoved cases respectively.
}
\description{
Plot SIR models on confirmed cases and removed cases.
}
\details{
Plot the models of infection and removed cases seperately with prediction and real data.
}
